% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_health_messages.R
\name{get_health_messages}
\alias{get_health_messages}
\title{Get health messages for AQHI+ risk categories}
\usage{
get_health_messages(risk_categories = NULL, language = "en")
}
\arguments{
\item{risk_categories}{(Optional).
A factor or character vector of AQHI risk categories (e.g. Low, Moderate, High, Very High).
If NULL (the default) all risk categories are returned, in \code{language}.}

\item{language}{(Optional).
A single character value indicating the language to use for risk levels and health messaging.
Must be either "en" (English) or "fr" (French). Not case sensitive.
Default is "en".}
}
\value{
A tibble of health messages (for at risk population and general population) for the provided risk categories.
}
\description{
Health messages for at risk population and general population are available for each AQHI health risk category (see \code{\link{get_risk_category}}).

English messages returned by this function:\tabular{llll}{
   Risk \tab AQHI \tab At Risk Population \tab General Population \cr
   Low \tab 1 - 3 \tab Enjoy your usual activities. \tab Ideal air quality for outdoor activities. \cr
   Moderate \tab 4 - 6 \tab Consider reducing or rescheduling activities outdoors if you experience symptoms. \tab No need to modify your usual activities unless you experience symptoms. \cr
   High \tab 7 - 10 \tab Reduce or reschedule activities outdoors. \tab Consider reducing or rescheduling activities outdoors if you experience symptoms. \cr
   Very High \tab >10 \tab Avoid strenuous activity outdoors. \tab Reduce or reschedule activities outdoors, especially if you experience symptoms. \cr
}


See \href{https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html}{Environment and Climate Change Canada's website} for more information.
}
\examples{
# Get health messages for all risk categories
get_health_messages()

# The same, but en Francais
get_health_messages(language = "fr")

# Get health messages for some observations
hourly_pm25_ugm3 <- sample(1:100, 50, replace = TRUE)
risk_categories <- hourly_pm25_ugm3 |>
  AQHI_plus(detailed = FALSE) |>
  get_risk_category()
risk_categories |> get_health_messages()
}
