% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAAQS.R
\name{CAAQS}
\alias{CAAQS}
\title{Assess the attainment of the Canadian Ambient Air Quality Standards (CAAQS)}
\usage{
CAAQS(
  dates,
  pm25_1hr_ugm3 = NULL,
  o3_1hr_ppb = NULL,
  no2_1hr_ppb = NULL,
  so2_1hr_ppb = NULL,
  min_completeness = 0.5
)
}
\arguments{
\item{dates}{Vector of hourly datetime values corresponding to observations. Date gaps will be filled automatically.}

\item{pm25_1hr_ugm3}{(Optional). Vector of hourly mean fine particulate matter (PM2.5) concentrations (ug/m^3).}

\item{o3_1hr_ppb}{(Optional). Vector of hourly mean ozone (O3) concentrations (ppb).}

\item{no2_1hr_ppb}{(Optional). Vector of hourly mean nitrogen dioxide (NO2) concentrations (ppb).}

\item{so2_1hr_ppb}{(Optional). Vector of hourly mean sulphur dioxide (SO2) concentrations (ppb).}

\item{min_completeness}{A single value from 0 to 1 indicating the required annual data completeness for a pollutant. Default is 0.5 (50 percent).}
}
\value{
a list of tibbles (data.frames), one tibble per pollutant provided with annual CAAQS metrics and management levels
}
\description{
The Canadian Ambient Air Quality Standards (CAAQS) are part of a collaborative national Air Quality Management System (AQMS), to better protect human health and the environment.
Standards at various averaging periods are defined for fine particulate matter (PM2.5), ozone (O3), nitrogen dioxide (NO2), and sulphur dioxide (SO2), and are typically updated ever 5 years.

Management levels (Green -> Yellow -> Orange -> Red) are defined for each pollutant standard.
A "Red" level indicates exceedance of the CAAQS and management plans are typically developed for regions at "Orange" or worse levels.
}
\examples{
obs <- data.frame(
  date = seq(
    lubridate::ymd_h("2020-01-01 00"),
    lubridate::ymd_h("2023-12-31 23"), "1 hours"
  ),
  pm25 = sample(1:150, 35064, TRUE), o3 = sample(1:150, 35064, TRUE),
  no2 = sample(1:150, 35064, TRUE), so2 = sample(1:150, 35064, TRUE)
)
CAAQS(
  dates = obs$date, pm25_1hr_ugm3 = obs$pm25,
  o3_1hr_ppb = obs$o3, no2_1hr_ppb = obs$no2, so2_1hr_ppb = obs$so2
)
}
\references{
\url{https://ccme.ca/en/air-quality-report}
}
\seealso{
Other Air Quality Standards: 
\code{\link{AQI}()}
}
\concept{Air Quality Standards}
\concept{Canadian Air Quality}
